requirements.production.txt
# Production Requirements for EESA Backend
# Optimized for PostgreSQL/Supabase deployment

# Core Django
Django==5.1.4
djangorestframework==3.15.2

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter for production
dj-database-url==2.2.0  # Database URL parsing

# Storage and Media
cloudinary==1.41.0
django-cloudinary-storage==0.3.0

# API and Filtering
django-filter==24.3
django-cors-headers==4.6.0

# Environment and Configuration
python-dotenv==1.0.1

# Production Server
gunicorn==23.0.0
whitenoise==6.8.2

# Caching (Optional but recommended)
django-redis==5.4.0
redis==5.2.0

# Security and Monitoring (Optional)
# sentry-sdk==2.18.0  # Uncomment for error monitoring

# Development and Testing (remove in production)
# django-debug-toolbar==4.4.6

# File Handling
Pillow==10.4.0

# Utils
pytz==2024.2
